fix: a generic version of the root modal
This can be moved to shared
